+++ config/routes.rb
@@ -73,6 +73,7 @@ Rails.application.routes.draw do
   get "import/export_status" => "import#export_status"
   get "import/download_backup" => "import#download_backup"
   get "import/report" => "import#report"
+  get "import/customer_report" => "import#customer_report"
   post "import/analyze" => "import#analyze"
   post "import/start" => "import#start"
   get "import/status" => "import#status"